fix minor bugs in edit category,  add category sort component, redying for sorting category component
