+++ src/app/components/main-pages/pages/category/category-crud/category-sort/category-sort.component.html
@@ -1,0 +1,5 @@
+<div cdkDropList cdkDropListOrientation="mixed" class="example-list" (cdkDropListDropped)="drop($event)">
+  @for (item of items; track item) {
+    <div class="example-box" cdkDrag>{{item}}</div>
+  }
+</div>
