--- src/app/components/main-pages/pages/category/category-sort/category-sort.component.html
@@ -1,1 +1,0 @@
-<p>category-sort works!</p>
