--- src/app/components/main-pages/pages/category/category-sort/category-sort.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-category-sort',
-  standalone: true,
-  imports: [],
-  templateUrl: './category-sort.component.html',
-  styleUrl: './category-sort.component.css'
-})
-export class CategorySortComponent {
-
-}
