+++ src/app/components/main-pages/pages/category/category.component.html
@@ -16,7 +16,7 @@
                 Export Categories
             </button>
         </mat-menu>
-        <button mat-stroked-button class="btnCategory">
+        <button mat-stroked-button class="btnCategory" [routerLink]="['sort-category']">
             <mat-icon>sort</mat-icon>
             Sort Categories
         </button>
